Assignment 1 - Vulnerable eHealth Application/app/api/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
Werkzeug==2.2.2
